refactor : 유저 지역 목록 DB에서 조회
